requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
alembic
pytest
pymysql
email-validator
bcrypt
jinja2
python-multipart